p3_miden_uni_stark

Trait StarkGenericConfig

pub trait StarkGenericConfig {
    type Pcs: Pcs<Self::Challenge, Self::Challenger>;
    type Challenge: ExtensionField<<<Self::Pcs as Pcs<Self::Challenge, Self::Challenger>>::Domain as PolynomialSpace>::Val>;
    type Challenger: FieldChallenger<<<Self::Pcs as Pcs<Self::Challenge, Self::Challenger>>::Domain as PolynomialSpace>::Val> + CanObserve<<Self::Pcs as Pcs<Self::Challenge, Self::Challenger>>::Commitment> + CanSample<Self::Challenge>;

    // Required methods
    fn pcs(&self) -> &Self::Pcs;
    fn initialise_challenger(&self) -> Self::Challenger;

    // Provided method
    fn is_zk(&self) -> usize { ... }
}

Required Associated Types

type Pcs: Pcs<Self::Challenge, Self::Challenger>

The Pcs implementation used to commit to trace polynomials.

type Challenge: ExtensionField<<<Self::Pcs as Pcs<Self::Challenge, Self::Challenger>>::Domain as PolynomialSpace>::Val>

The ExtensionField from which most random challenges are drawn.

type Challenger: FieldChallenger<<<Self::Pcs as Pcs<Self::Challenge, Self::Challenger>>::Domain as PolynomialSpace>::Val> + CanObserve<<Self::Pcs as Pcs<Self::Challenge, Self::Challenger>>::Commitment> + CanSample<Self::Challenge>

The FieldChallenger (Fiat-Shamir) implementation used.

Required Methods

fn pcs(&self) -> &Self::Pcs

Get a reference to the Pcs used by this proof configuration.

fn initialise_challenger(&self) -> Self::Challenger

Get an initialisation of the FieldChallenger used by this proof configuration.

Provided Methods

fn is_zk(&self) -> usize

Returns 1 if the Pcs is zero-knowledge, 0 otherwise.